Fix TS modules not being resolved
There is a discrepancy between the ways typescript-language-server and
webpack loaders resolve modules. In IDE, it correctly resolves the
module WITHOUT .ts extension, but fails at compile step in webpack. The
solution is to provide extensions automatically resolved by webpack.

Note, that it's needed to explicitly add extensions resolved by default
in case of resolve.extensions absence, that is '.js', '.json'